refactored common aws functionality into Rubber::Cloud::Aws::Base
